Client Class Refactor
Completed a major overhaul of the client class in preparation for implementing artificial intelligence for single player (#4). There is now a base AbstractClient class which is extended by the player's view (GUIClient Class). The AI will also extend the AbstractClient class.

Revised the SocketIODriver code to handle the socket in addition to the I/O streams. This will allow the socket I/O initialization to be encapsulated from the GUIClient and AIClient implementations.

Added a main class for more obvious entry to the application.